+++ account/go.mod
@@ -1,0 +1,17 @@
+module account
+
+go 1.22.0
+
+require (
+	github.com/golang/protobuf v1.5.3
+	github.com/google/uuid v1.6.0
+	google.golang.org/grpc v1.62.1
+	google.golang.org/protobuf v1.32.0
+)
+
+require (
+	golang.org/x/net v0.20.0 // indirect
+	golang.org/x/sys v0.16.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20240123012728-ef4313101c80 // indirect
+)
